Move group compression to server side
